requirements.txt
kaggle
zipfile
cv2
numpy
pandas
sklearn
tensorflow
matplotlib
flask
Pillow